requirements.txt
streamlit>=1.28.0
langgraph>=0.0.40
langchain-openai>=0.1.0
langchain-community>=0.2.0
tavily-python>=0.3.0
openai>=1.0.0
asyncio
typing-extensions>=4.5.0